+++ packages/feedback-components/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file. The format
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.3] - 2025-07-16
+
+- Graph node clicking logic same as graph and tree
+- "Show matches" switch to displayed matchs on tag hover
+
 ## [1.1.2] - 2025-07-10
 
 - Increased the default line spacing in the paragraph text (made it configurable